Refactor: AivmManager のメタデータ管理部分を AivmInfosRepository に切り出し、前回起動時の情報をキャッシュしてエンジン起動を大幅に高速化する
リファクタの手間にあっていたかは微妙（もはやリファクタがメイン）だが、肥大化しすぎていた実装をうまく切り出せて満足ではある
